Unlink symlinks to directories
If the vendoring directory contains a symlink to directory, remove the
symlink instead of nuking the linked directory.

+++ src/vendoring/utils.py
@@ -11,7 +11,7 @@ def remove_all(items_to_cleanup: Iterable[Path]) -> None:
     for item in items_to_cleanup:
         if not item.exists():
             continue
-        if item.is_dir():
+        if item.is_dir() and not item.is_symlink():
             shutil.rmtree(str(item))
         else:
             item.unlink()
